rearrange functions related to process execution in the console application
